+++ rhel7/build-zato.sh
@@ -136,9 +136,12 @@ function checkout_zato {
 function install_zato {
     cd $ZATO_TARGET_DIR/code
     sed -i -e 's|pg8000==1.13.1|pg8000==1.12.5|' -e 's|pyasn1==0.4.5|pyasn1==0.4.8|' requirements.txt
-    sed -i -e 's|bzr==2.6.0|bzr==2.7.0|' _req_py27.txt
+
     ./install.sh -p ${PY_BINARY}
-    run_tests_zato || exit 1
+    if [[ "${SKIP_TESTS:-n}" == "y" ]]; then
+        run_tests_zato || exit 1
+    fi
+
     find $ZATO_TARGET_DIR/. -name *.pyc -exec rm -f {} \;
     find $ZATO_TARGET_DIR/. ! -perm /004 -exec chmod 644 {} \;
     [[ -f ./code/hotfixman.sh ]] && rm -f ./code/hotfixman.sh
