+++ rhel8/build-zato.sh
@@ -110,7 +110,9 @@ function install_zato {
     sudo ${INSTALL_CMD} config-manager --set-enabled PowerTools
     
     ./install.sh -p ${PY_BINARY}
-    run_tests_zato || exit 1
+    if [[ "${SKIP_TESTS:-n}" == "y" ]]; then
+        run_tests_zato || exit 1
+    fi
 
     find $ZATO_TARGET_DIR/. -name *.pyc -exec rm -f {} \;
     find $ZATO_TARGET_DIR/. ! -perm /004 -exec chmod 644 {} \;
